+++ src/app/data_harvester/battery_harvester.rs
@@ -1,20 +1,19 @@
 use battery::{
-    units::{electric_potential::volt, power::watt, ratio::percent, time::second, Time},
+    units::{power::watt, ratio::percent, time::second, Time},
     Battery, Manager,
 };
 
 #[derive(Debug, Clone)]
 pub struct BatteryHarvest {
-    charge_percent: f64,
-    secs_until_full: Option<f64>,
-    secs_until_empty: Option<f64>,
-    power_consumption_rate_watts: f64,
-    voltage: f64,
+    pub charge_percent: u64,
+    pub secs_until_full: Option<i64>,
+    pub secs_until_empty: Option<i64>,
+    pub power_consumption_rate_watts: f64,
 }
 
-fn convert_optional_time_to_optional_seconds(optional_time: Option<Time>) -> Option<f64> {
+fn convert_optional_time_to_optional_seconds(optional_time: Option<Time>) -> Option<i64> {
     if let Some(time) = optional_time {
-        Some(f64::from(time.get::<second>()))
+        Some(f64::from(time.get::<second>()) as i64)
     } else {
         None
     }
@@ -32,8 +31,7 @@ pub fn refresh_batteries(manager: &Manager, batteries: &mut [Battery]) -> Vec<Ba
                     secs_until_empty: convert_optional_time_to_optional_seconds(
                         battery.time_to_empty(),
                     ),
-                    charge_percent: f64::from(battery.state_of_charge().get::<percent>()),
-                    voltage: f64::from(battery.voltage().get::<volt>()),
+                    charge_percent: f64::from(battery.state_of_charge().get::<percent>()) as u64,
                     power_consumption_rate_watts: f64::from(battery.energy_rate().get::<watt>()),
                 })
             } else {